+++ PRN222.Assignment.FPTURoomBooking.Services/Services/DepartmentService.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq.Expressions;
 using Mapster;
+using Microsoft.EntityFrameworkCore;
 using PRN222.Assignment.FPTURoomBooking.Repositories.Models;
 using PRN222.Assignment.FPTURoomBooking.Repositories.UnitOfWork;
 using PRN222.Assignment.FPTURoomBooking.Services.Models.Department;
@@ -27,11 +28,19 @@ namespace PRN222.Assignment.FPTURoomBooking.Services.Services
 
         public async Task<Result> DeleteAsync(Guid id)
         {
-            var entity = await _unitOfWork.DepartmentRepository.GetByIdAsync(id);
+            var entity = await _unitOfWork.DepartmentRepository.GetQueryable()
+                .Include(x => x.Accounts)
+                .FirstOrDefaultAsync(x => x.Id == id);
             if (entity == null)
             {
                 return Result.Failure("Department not found");
             }
+
+            if (entity.Accounts.Any())
+            {
+                return Result.Failure("Cannot delete department with associated accounts");
+            }
+
             _unitOfWork.DepartmentRepository.Remove(entity);
             await _unitOfWork.SaveChangesAsync();
             return Result.Success();
@@ -39,12 +48,11 @@ namespace PRN222.Assignment.FPTURoomBooking.Services.Services
 
         public async Task<Result<DepartmentModel>> GetAsync(Guid id)
         {
-            var entity = await _unitOfWork.DepartmentRepository.GetByIdAsync(id);
-            if (entity == null)
-            {
-                return Result<DepartmentModel>.Failure("Department not found");
-            }
-            return entity.Adapt<DepartmentModel>();
+            var entity = await _unitOfWork.DepartmentRepository.GetQueryable()
+                .Include(x => x.Campus)
+                .ProjectToType<DepartmentModel>()
+                .FirstOrDefaultAsync(x => x.Id == id);
+            return entity ?? Result<DepartmentModel>.Failure("Department not found");
         }
 
         public async Task<Result<PaginationResult<DepartmentModel>>> GetPagedAsync(GetDepartmentModel model)
@@ -56,7 +64,7 @@ namespace PRN222.Assignment.FPTURoomBooking.Services.Services
             {
                 filter = filter.CombineAndAlsoExpressions(x => true);
             }
-            
+
             if (!model.CampusId.IsNullOrGuidEmpty())
             {
                 filter = filter.CombineAndAlsoExpressions(x => x.CampusId == model.CampusId);
